moved rendergoblins to utils and tested it
