+++ test/example.test.js
@@ -7,15 +7,11 @@ const test = QUnit.test;
 test('time to test a function', (expect) => {
     //Arrange
     // Set up your arguments and expectations
-    const expected = `<div class="goblin">
-<p>Ross from friends</p>
-<p id="face1">😈</p>
-<p id="hp1">666</p>
-</div>`;
+    const expected = `<div class="goblin"><p>Randy Johnson</p><p>&gt;:)</p><p id="GoblinHP1">8</p></div>`;
     
     //Act 
     // Call the function you're testing and set the result to a const
-    const actual = renderGoblin({ name: `Ross from friends`, id: 1, hp: `666` });
+    const actual = renderGoblin({ name: `Randy Johnson`, id: 1, hp: `8` });
 
     //Expect
     // Make assertions about what is expected versus the actual result
